# Performance optimizations for train.py on CUDA GPUs