--- a/utils/loss.py
+++ b/utils/loss.py
@@ -14,6 +14,12 @@ def smooth_BCE(eps=0.1):  # https://github.com/ultralytics/yolov3/issues/238#iss
     # return positive, negative label smoothing BCE targets
     return 1.0 - 0.5 * eps, 0.5 * eps
 
+def reg_loss(pxy, pwh, tbox_i, anchors_i):
+    pxy = pxy.sigmoid() * 2 - 0.5
+    pwh = (pwh.sigmoid() * 2) ** 2 * anchors_i
+    pbox = torch.cat((pxy, pwh), 1)  # predicted box
+    iou = bbox_iou(pbox, tbox_i, CIoU=True).squeeze()  # iou(prediction, target)
+    return iou
 
 class BCEBlurWithLogitsLoss(nn.Module):
     # BCEwithLogitLoss() with reduced missing label effects.
@@ -118,6 +124,28 @@ class ComputeLoss:
         self.anchors = m.anchors
         self.device = device
 
+        if self.device.type == 'cuda':
+
+            self.num_max_pred_pad = 6400
+            _pxy = torch.empty([self.num_max_pred_pad, 2], 
+                    dtype=torch.float16, device='cuda', 
+                    requires_grad=True)
+            _pwh = torch.empty([self.num_max_pred_pad, 2], 
+                    dtype=torch.float32, device='cuda', 
+                    requires_grad=True)
+            _tbox_i = torch.empty([self.num_max_pred_pad, 4], 
+                    dtype=torch.float32, device='cuda', 
+                    requires_grad=False)
+            _anchors_i = torch.empty([self.num_max_pred_pad, 2], 
+                    dtype=torch.float32, device='cuda', 
+                    requires_grad=False)
+            
+            self.reg_loss= torch.jit.trace(reg_loss, (_pxy, _pwh, _tbox_i, _anchors_i)) 
+            del _pxy, _pwh, _tbox_i
+        else:
+            self.reg_loss = reg_loss
+            self.num_max_pred_pad = 0
+
     def __call__(self, p, targets):  # predictions, targets
         lcls = torch.zeros(1, device=self.device)  # class loss
         lbox = torch.zeros(1, device=self.device)  # box loss
@@ -134,11 +162,17 @@ class ComputeLoss:
                 # pxy, pwh, _, pcls = pi[b, a, gj, gi].tensor_split((2, 4, 5), dim=1)  # faster, requires torch 1.8.0
                 pxy, pwh, _, pcls = pi[b, a, gj, gi].split((2, 2, 1, self.nc), 1)  # target-subset of predictions
 
+                # Pad
+                p_size = pxy.size(0)
+                pad_size = max(0, self.num_max_pred_pad - p_size)
+                pxy = torch.nn.functional.pad(pxy, (0, 0, 0, pad_size))
+                pwh = torch.nn.functional.pad(pwh, (0, 0, 0, pad_size))
+                tbox_i= torch.nn.functional.pad(tbox[i], (0, 0, 0, pad_size))
+                anchors_i= torch.nn.functional.pad(anchors[i], (0, 0, 0, pad_size))
+
                 # Regression
-                pxy = pxy.sigmoid() * 2 - 0.5
-                pwh = (pwh.sigmoid() * 2) ** 2 * anchors[i]
-                pbox = torch.cat((pxy, pwh), 1)  # predicted box
-                iou = bbox_iou(pbox, tbox[i], CIoU=True).squeeze()  # iou(prediction, target)
+                iou = self.reg_loss(pxy, pwh, tbox_i, anchors_i)
+                iou = iou[:p_size]
                 lbox += (1.0 - iou).mean()  # iou loss
 
                 # Objectness
@@ -174,6 +208,7 @@ class ComputeLoss:
 
         return (lbox + lobj + lcls) * bs, torch.cat((lbox, lobj, lcls)).detach()
 
+    
     def build_targets(self, p, targets):
         # Build targets for compute_loss(), input targets(image,class,x,y,w,h)
         na, nt = self.na, targets.shape[0]  # number of anchors, targets