<!--
Thank you for submitting a YOLOv5 🚀 Pull Request! We want to make contributing to YOLOv5 as easy and transparent as possible. A few tips to get you started:

- Search existing YOLOv5 [PRs](https://github.com/ultralytics/yolov5/pull) to see if a similar PR already exists.
- Link this PR to a YOLOv5 [issue](https://github.com/ultralytics/yolov5/issues) to help us understand what bug fix or feature is being implemented.
- Provide before and after profiling/inference/training results to help us quantify the improvement your PR provides (if applicable).

Please see our ✅ [Contributing Guide](https://github.com/ultralytics/yolov5/blob/master/CONTRIBUTING.md) for more details.
-->

In the process of trying out yolov5, I discovered some system-level optimization opportunities. These opportunities mostly work with NVIDIA GPUs. The optimizations gave me around 20-30% performance improvement on mscoco dataset, I thought I'd share these with the community and seek feedback (if it can be merged).

1. Use NVIDIA's fused optimizer when available (this requires a manual installation of the apex module [github](https://github.com/NVIDIA/apex). I am struggling to find a good place for documenting the installation steps, as it cannot be simply put into requirements.txt
2. use torch.jit.trace for part of the bbox regression loss computation as they involve many element-wise operations. The benefit of this mostly depends on the jit backend (nvfuser) and the workload type. Also, some padding is added to reduce the shape dynamism for jit to get its best performance.
3. use channel last on the backbone when amp is enabled 

training log before the change (on `master`)
```
yolov5$ python train.py --data data/coco.yaml --weights '' --cfg yolov5m.yaml --noval --epochs 2 --batch-size 64 --optimizer Adam
train: weights=, cfg=yolov5m.yaml, data=data/coco.yaml, hyp=data/hyps/hyp.scratch-low.yaml, epochs=2, batch_size=64, imgsz=640, rect=False, resume=False, nosave=False, noval=True, noautoanchor=False, noplots=False, evolve=None, bucket=, cache=None, image_weights=False, device=, multi_scale=False, single_cls=False, optimizer=Adam, sync_bn=False, workers=8, project=runs/train, name=exp, exist_ok=False, quad=False, cos_lr=False, label_smoothing=0.0, patience=100, freeze=[0], save_period=-1, seed=0, local_rank=-1, entity=None, upload_dataset=False, bbox_interval=-1, artifact_alias=latest
github: ⚠️ YOLOv5 is out of date by 3 commits. Use `git pull ultralytics master` or `git clone https://github.com/ultralytics/yolov5` to update.
YOLOv5 🚀 v7.0-5-gbfa1f23 Python-3.8.15 torch-1.13.0+cu117 CUDA:0 (NVIDIA GeForce RTX 3090, 24268MiB)

hyperparameters: lr0=0.01, lrf=0.01, momentum=0.937, weight_decay=0.0005, warmup_epochs=3.0, warmup_momentum=0.8, warmup_bias_lr=0.1, box=0.05, cls=0.5, cls_pw=1.0, obj=1.0, obj_pw=1.0, iou_t=0.2, anchor_t=4.0, fl_gamma=0.0, hsv_h=0.015, hsv_s=0.7, hsv_v=0.4, degrees=0.0, translate=0.1, scale=0.5, shear=0.0, perspective=0.0, flipud=0.0, fliplr=0.5, mosaic=1.0, mixup=0.0, copy_paste=0.0
ClearML: run 'pip install clearml' to automatically track, visualize and remotely train YOLOv5 🚀 in ClearML
Comet: run 'pip install comet_ml' to automatically track and visualize YOLOv5 🚀 runs in Comet
TensorBoard: Start with 'tensorboard --logdir runs/train', view at http://localhost:6006/

                 from  n    params  module                                  arguments                     
  0                -1  1      5280  models.common.Conv                      [3, 48, 6, 2, 2]              
  1                -1  1     41664  models.common.Conv                      [48, 96, 3, 2]                
  2                -1  2     65280  models.common.C3                        [96, 96, 2]                   
  3                -1  1    166272  models.common.Conv                      [96, 192, 3, 2]               
  4                -1  4    444672  models.common.C3                        [192, 192, 4]                 
  5                -1  1    664320  models.common.Conv                      [192, 384, 3, 2]              
  6                -1  6   2512896  models.common.C3                        [384, 384, 6]                 
  7                -1  1   2655744  models.common.Conv                      [384, 768, 3, 2]              
  8                -1  2   4134912  models.common.C3                        [768, 768, 2]                 
  9                -1  1   1476864  models.common.SPPF                      [768, 768, 5]                 
 10                -1  1    295680  models.common.Conv                      [768, 384, 1, 1]              
 11                -1  1         0  torch.nn.modules.upsampling.Upsample    [None, 2, 'nearest']          
 12           [-1, 6]  1         0  models.common.Concat                    [1]                           
 13                -1  2   1182720  models.common.C3                        [768, 384, 2, False]          
 14                -1  1     74112  models.common.Conv                      [384, 192, 1, 1]              
 15                -1  1         0  torch.nn.modules.upsampling.Upsample    [None, 2, 'nearest']          
 16           [-1, 4]  1         0  models.common.Concat                    [1]                           
 17                -1  2    296448  models.common.C3                        [384, 192, 2, False]          
 18                -1  1    332160  models.common.Conv                      [192, 192, 3, 2]              
 19          [-1, 14]  1         0  models.common.Concat                    [1]                           
 20                -1  2   1035264  models.common.C3                        [384, 384, 2, False]          
 21                -1  1   1327872  models.common.Conv                      [384, 384, 3, 2]              
 22          [-1, 10]  1         0  models.common.Concat                    [1]                           
 23                -1  2   4134912  models.common.C3                        [768, 768, 2, False]          
 24      [17, 20, 23]  1    343485  models.yolo.Detect                      [80, [[10, 13, 16, 30, 33, 23], [30, 61, 62, 45, 59, 119], [116, 90, 156, 198, 373, 326]], [192, 384, 768]]
YOLOv5m summary: 291 layers, 21190557 parameters, 21190557 gradients, 49.2 GFLOPs

AMP: checks passed ✅
optimizer: Adam(lr=0.01) with parameter groups 79 weight(decay=0.0), 82 weight(decay=0.0005), 82 bias
train: Scanning /home/xinli/projects/datasets/coco/train2017.cache... 117266 images, 1021 backgrounds, 0 corrupt: 100%|██████████| 118287/118287 00:00
train: WARNING ⚠️ /home/xinli/projects/datasets/coco/images/train2017/000000099844.jpg: 2 duplicate labels removed
train: WARNING ⚠️ /home/xinli/projects/datasets/coco/images/train2017/000000201706.jpg: 1 duplicate labels removed
train: WARNING ⚠️ /home/xinli/projects/datasets/coco/images/train2017/000000214087.jpg: 1 duplicate labels removed
train: WARNING ⚠️ /home/xinli/projects/datasets/coco/images/train2017/000000522365.jpg: 1 duplicate labels removed
val: Scanning /home/xinli/projects/datasets/coco/val2017.cache... 4952 images, 48 backgrounds, 0 corrupt: 100%|██████████| 5000/5000 00:00

AutoAnchor: 4.45 anchors/target, 0.995 Best Possible Recall (BPR). Current anchors are a good fit to dataset ✅
Plotting labels to runs/train/exp37/labels.jpg... 
Image sizes 640 train, 640 val
Using 8 dataloader workers
Logging results to runs/train/exp37
Starting training for 2 epochs...

      Epoch    GPU_mem   box_loss   obj_loss   cls_loss  Instances       Size
        0/1      22.4G    0.07281    0.08384    0.06359        199        640: 100%|██████████| 1849/1849 13:15

      Epoch    GPU_mem   box_loss   obj_loss   cls_loss  Instances       Size
        1/1      22.4G    0.06403    0.08171    0.05178        168        640: 100%|██████████| 1849/1849 13:13 
                 Class     Images  Instances          P          R      mAP50   mAP50-95: 100%|██████████| 40/40 00:24
                   all       5000      36335      0.301      0.127     0.0926     0.0434

2 epochs completed in 0.449 hours.
Optimizer stripped from runs/train/exp37/weights/last.pt, 42.7MB
Optimizer stripped from runs/train/exp37/weights/best.pt, 42.7MB

Validating runs/train/exp37/weights/best.pt...
Fusing layers... 
YOLOv5m summary: 212 layers, 21172173 parameters, 0 gradients, 48.9 GFLOPs
                 Class     Images  Instances          P          R      mAP50   mAP50-95: 100%|██████████| 40/40 00:31
                   all       5000      36335      0.298      0.126     0.0897     0.0428
                person       5000      10777      0.222      0.635      0.477      0.224
               bicycle       5000        314      0.247     0.0723     0.0539     0.0176
                   car       5000       1918      0.118       0.46      0.269      0.117
            motorcycle       5000        367      0.156      0.177     0.0948      0.032
              airplane       5000        143      0.217      0.182       0.15     0.0675
                   bus       5000        283      0.403      0.158      0.196      0.104
                 train       5000        190      0.142      0.116     0.0624      0.022
                 truck       5000        414      0.123      0.241     0.0803     0.0396
                  boat       5000        424      0.119     0.0849     0.0439     0.0137
         traffic light       5000        634      0.297      0.131       0.11     0.0498
          fire hydrant       5000        101      0.151      0.139      0.113     0.0718
             stop sign       5000         75      0.767      0.227      0.294      0.208
         parking meter       5000         60          1          0     0.0448     0.0243
                 bench       5000        411      0.293     0.0073     0.0132    0.00619
                  bird       5000        427      0.109     0.0328     0.0252     0.0111
                   cat       5000        202      0.441     0.0693      0.122     0.0446
                   dog       5000        218      0.109     0.0321     0.0634     0.0307
                 horse       5000        272      0.197      0.188      0.112     0.0425
                 sheep       5000        354      0.153      0.331      0.109     0.0529
                   cow       5000        372      0.215      0.317      0.181     0.0919
              elephant       5000        252      0.266      0.171      0.121     0.0585
                  bear       5000         71      0.478      0.052     0.0996      0.053
                 zebra       5000        266      0.306      0.583      0.439      0.177
               giraffe       5000        232      0.389      0.397      0.358      0.144
              backpack       5000        371      0.205     0.0167     0.0151    0.00652
              umbrella       5000        407      0.318     0.0934     0.0803     0.0334
               handbag       5000        540     0.0551      0.013     0.0079    0.00335
                   tie       5000        252     0.0782     0.0357     0.0155    0.00466
              suitcase       5000        299      0.182     0.0234     0.0281     0.0126
               frisbee       5000        115      0.252      0.217      0.129     0.0716
                  skis       5000        241     0.0577      0.141     0.0468     0.0179
             snowboard       5000         69     0.0222     0.0145    0.00538    0.00217
           sports ball       5000        260      0.154      0.365      0.225      0.133
                  kite       5000        327      0.264      0.309      0.218       0.11
          baseball bat       5000        145          1    0.00944     0.0475     0.0206
        baseball glove       5000        148      0.126      0.196     0.0714     0.0383
            skateboard       5000        179      0.168      0.101     0.0502     0.0173
             surfboard       5000        267      0.128     0.0899     0.0316    0.00978
         tennis racket       5000        225      0.291      0.111      0.095     0.0414
                bottle       5000       1013      0.164      0.224      0.121     0.0522
            wine glass       5000        341      0.145    0.00587     0.0242     0.0069
                   cup       5000        895      0.261      0.201      0.149     0.0743
                  fork       5000        215          1          0    0.00768    0.00266
                 knife       5000        325          1          0     0.0054    0.00264
                 spoon       5000        253          1          0     0.0111    0.00427
                  bowl       5000        623      0.171      0.128     0.0805     0.0456
                banana       5000        370       0.22     0.0378     0.0341     0.0112
                 apple       5000        236          0          0     0.0129    0.00794
              sandwich       5000        177     0.0363    0.00565     0.0417      0.018
                orange       5000        285          0          0     0.0422     0.0279
              broccoli       5000        312      0.221      0.205      0.118     0.0428
                carrot       5000        365      0.405    0.00274     0.0448     0.0216
               hot dog       5000        125          0          0     0.0124     0.0054
                 pizza       5000        284      0.458      0.197      0.204     0.0986
                 donut       5000        328      0.202     0.0427     0.0495     0.0299
                  cake       5000        310       0.05     0.0226     0.0255     0.0119
                 chair       5000       1771      0.185     0.0418     0.0408     0.0169
                 couch       5000        261      0.359     0.0345     0.0794     0.0323
          potted plant       5000        342      0.131      0.167     0.0653     0.0257
                   bed       5000        163      0.315     0.0307     0.0454     0.0131
          dining table       5000        695      0.355     0.0261     0.0571     0.0175
                toilet       5000        179      0.113      0.397      0.187      0.105
                    tv       5000        288     0.0996      0.441      0.195      0.107
                laptop       5000        231      0.205     0.0649     0.0393     0.0148
                 mouse       5000        106     0.0554      0.387     0.0997      0.065
                remote       5000        283      0.234    0.00353     0.0194    0.00622
              keyboard       5000        153      0.295      0.085     0.0797     0.0346
            cell phone       5000        262      0.072     0.0116     0.0134    0.00757
             microwave       5000         55      0.138     0.0364     0.0362     0.0188
                  oven       5000        143          1          0      0.043     0.0156
               toaster       5000          9          1          0          0          0
                  sink       5000        225      0.137      0.204      0.089     0.0476
          refrigerator       5000        126      0.151     0.0397     0.0302     0.0174
                  book       5000       1129      0.081     0.0523     0.0204    0.00586
                 clock       5000        267      0.265      0.378        0.3      0.161
                  vase       5000        274     0.0748     0.0328     0.0223     0.0116
              scissors       5000         36          1          0          0          0
            teddy bear       5000        190          1          0      0.029     0.0142
            hair drier       5000         11          0          0          0          0
            toothbrush       5000         57          1          0    0.00118   0.000355

Evaluating pycocotools mAP... saving runs/train/exp37/_predictions.json...
loading annotations into memory...
Done (t=0.47s)
creating index...
index created!
Loading and preparing results...
DONE (t=3.70s)
creating index...
index created!
Running per image evaluation...
Evaluate annotation type *bbox*
DONE (t=24.20s).
Accumulating evaluation results...
DONE (t=8.86s).
 Average Precision  (AP) @[ IoU=0.50:0.95 | area=   all | maxDets=100 ] = 0.043
 Average Precision  (AP) @[ IoU=0.50      | area=   all | maxDets=100 ] = 0.091
 Average Precision  (AP) @[ IoU=0.75      | area=   all | maxDets=100 ] = 0.037
 Average Precision  (AP) @[ IoU=0.50:0.95 | area= small | maxDets=100 ] = 0.037
 Average Precision  (AP) @[ IoU=0.50:0.95 | area=medium | maxDets=100 ] = 0.059
 Average Precision  (AP) @[ IoU=0.50:0.95 | area= large | maxDets=100 ] = 0.042
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=   all | maxDets=  1 ] = 0.093
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=   all | maxDets= 10 ] = 0.201
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=   all | maxDets=100 ] = 0.233
 Average Recall     (AR) @[ IoU=0.50:0.95 | area= small | maxDets=100 ] = 0.122
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=medium | maxDets=100 ] = 0.276
 Average Recall     (AR) @[ IoU=0.50:0.95 | area= large | maxDets=100 ] = 0.256
Results saved to runs/train/exp37
```

Training log after the change
```
yolov5$ python train.py --data data/coco.yaml --weights '' --cfg yolov5m.yaml --noval --epochs 2 --batch-size 64 --optimizer Adam
train: weights=, cfg=yolov5m.yaml, data=data/coco.yaml, hyp=data/hyps/hyp.scratch-low.yaml, epochs=2, batch_size=64, imgsz=640, rect=False, resume=False, nosave=False, noval=True, noautoanchor=False, noplots=False, evolve=None, bucket=, cache=None, image_weights=False, device=, multi_scale=False, single_cls=False, optimizer=Adam, sync_bn=False, workers=8, project=runs/train, name=exp, exist_ok=False, quad=False, cos_lr=False, label_smoothing=0.0, patience=100, freeze=[0], save_period=-1, seed=0, local_rank=-1, entity=None, upload_dataset=False, bbox_interval=-1, artifact_alias=latest
github: ⚠️ YOLOv5 is out of date by 3 commits. Use `git pull ultralytics master` or `git clone https://github.com/ultralytics/yolov5` to update.
YOLOv5 🚀 v7.0-8-g05468d1 Python-3.8.15 torch-1.13.0+cu117 CUDA:0 (NVIDIA GeForce RTX 3090, 24268MiB)

hyperparameters: lr0=0.01, lrf=0.01, momentum=0.937, weight_decay=0.0005, warmup_epochs=3.0, warmup_momentum=0.8, warmup_bias_lr=0.1, box=0.05, cls=0.5, cls_pw=1.0, obj=1.0, obj_pw=1.0, iou_t=0.2, anchor_t=4.0, fl_gamma=0.0, hsv_h=0.015, hsv_s=0.7, hsv_v=0.4, degrees=0.0, translate=0.1, scale=0.5, shear=0.0, perspective=0.0, flipud=0.0, fliplr=0.5, mosaic=1.0, mixup=0.0, copy_paste=0.0
ClearML: run 'pip install clearml' to automatically track, visualize and remotely train YOLOv5 🚀 in ClearML
Comet: run 'pip install comet_ml' to automatically track and visualize YOLOv5 🚀 runs in Comet
TensorBoard: Start with 'tensorboard --logdir runs/train', view at http://localhost:6006/

                 from  n    params  module                                  arguments                     
  0                -1  1      5280  models.common.Conv                      [3, 48, 6, 2, 2]              
  1                -1  1     41664  models.common.Conv                      [48, 96, 3, 2]                
  2                -1  2     65280  models.common.C3                        [96, 96, 2]                   
  3                -1  1    166272  models.common.Conv                      [96, 192, 3, 2]               
  4                -1  4    444672  models.common.C3                        [192, 192, 4]                 
  5                -1  1    664320  models.common.Conv                      [192, 384, 3, 2]              
  6                -1  6   2512896  models.common.C3                        [384, 384, 6]                 
  7                -1  1   2655744  models.common.Conv                      [384, 768, 3, 2]              
…