@AyushExel 2022-11-28
@xinli-centml I'll chime in here because the speedups look impressive. I think the path forward here should be to make these requirements optional as some changes require manual installation for now. It'll require a short guide in the tutorials section of readme like- "Speed-up Training: apex & JIT optimizations".