@glenn-jocher 2022-11-30
@xinli-centml this is really interesting, thanks for the PR! Is there a way to embed the EMA update improvements in the `ema.update()` method itself?