@xinli-centml 2022-11-30
Hi @glenn-jocher , Yes, I think that can be done and would be a better implementation, although we still need to add one extra synchronize() before `optimizer.step` to ensure ema update is finished before we update the weights again. 

I'll make the change