@xinli-centml 2022-12-13
Hi I think this PR is ready for a review, github actions was passing from the lastest commit (before merging main)

Re: ema to a different stream: I removed this optimization mainly because the performance improvements are small and are also dependent on batch size and hardware resource availability. However, it introduces quite a few changes to the main training loop which would impact a lot of people, and reduce code quality, maintanability, etc. I fianlly decided that the gain is not worth it.

The other changes: channels last, fused optimizer and jit are mostly safe changes. For users using NVIDIA's NGC container, Apex is automatically installed and requires no extra setup.